update cpp development environment

+++ workbench_cpp/original.cpp
@@ -5,8 +5,8 @@ using ll = long long;
 #define rep(i, n) for (int i = 0; i < int(n); i++)
 #define repi(i, a, b) for (int i = int(a); i < int(b); ++i)
 
-long long n;
-vector<long long> a(pow(10, 6));
+ll n;
+vector<ll> a(pow(10, 6));
 
 int main() {
     cin >> n;
